[TASK] Use SVG icons in EXT:redirects
The module icon is in SVG format with a background color from
the TYPO3 icon set. A similar icon for the record type has been
added too. The old PNG icons were removed.
In a future clean-up of the TYPO3 icon set these will be added to the
collection.

Resolves: #83653
Releases: master
Change-Id: Iec1a8b0344b91c43e501efa4deea9584ddf8b561
Reviewed-on: https://review.typo3.org/55456
Reviewed-by: Reiner Teubner <rteubner@me.com>
Tested-by: Reiner Teubner <rteubner@me.com>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Frank Naegler <frank.naegler@typo3.org>
Tested-by: Frank Naegler <frank.naegler@typo3.org>