[FEATURE] Add feature switch to reorder redirect middlewares
A new feature switch has been introduced to change the ordering
of the redirect middleware and the base redirect middleware.

If enabled, the redirect is executed before the
base-redirect-resolver to ensure configured redirects are hit first.

To not break the current behaviour the new order is not
applied by default.

Resolves: #88902
Releases: master
Change-Id: I3ba732f4611c03f589c10bf74ddcb61deefe8816
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61426
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>