[BUGFIX] Allow redirects to specific records
Creating links to specific records (LinkHandler functionality) does require
a global TSFE object. In addition, calling LinkBuilder directly will
result in a fatal error when linking e.g. to EXT:news records.

Currently, this is not possible as the RedirectService is not prepared for this.

The patch modifies the places accordingly.

Resolves: #87574
Releases: master, 9.5
Change-Id: Iebda76bad0a2681df30f2ea4c42cfd3b5cd4a39f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62128
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>