[TASK] Add missing tooltips in various modules
This (re)adds various tooltips through the backend to
further increase the accessibility. Especially for buttons
that have a different meaning than normal in a particular context.

In most cases, they were simply missed but some of them referenced
to not longer existing locallang keys or to locallang keys of
another sysext which is no dependency of the one where it is referenced.

The tooltips are (re)added at following places:
- Creating new content element in ext:core new record view
- Multiple buttons in ext:core element information modal
- Multiple buttons in ext:beuser listing and compare view
- Multiple buttons in ext:form form management and form editor view
- Multiple buttons in ext:info pagetree overview
- New content element record button in ext:recordlist listing view
- View redirect button in ext:redirect listing view

Resolves: #89566
Releases: master
Change-Id: I66cd16413b0e4f188627df06f233a0bba73c3056
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/62175
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Michael Telgkamp <michael.telgkamp@mindscreen.de>
Reviewed-by: Susanne Moog <look@susi.dev>