[BUGFIX] Do not flush redirect cache when re-building
Flushing the redirect cache before rebuilding
is not necessary and may result in dead-locks
if multiple processes try to do the same.

As there is only a single entry for redirects
which gets overwritten when the rebuilding is
finished, there is also no need to tag the entry
as the identifier is enough.

Resolves: #90069
Releases: master, 9.5
Change-Id: I2bc7c5ce9d0f4a37a3b1c9f51df34de556b2f35f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63743
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>