[TASK] Replace <style> for compliance with CSP header in svgs
SVG with inline styles are not loaded when the Content-Security-Policy
header contains the widely used setting "style-src 'self';" because
the browser has to reject interpreting the style attribute.
In this case a black rectangle is displayed.
Using attributes instead of styles is compliant with CSP "style-src
'self';" and the svgs can be displayed.

Releases: master, 10.4
Resolves: #91216
Change-Id: I1679a2004f1fc868231318249a267cbba62c540f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/64331
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>