[TASK] Unify backend filter styling
Improves the overall look and feel of the filters in
various backend modules, thanks to the new bootstrap v5
grid. We can therefore now rely on flex-form instead of
float technology. By using the `g-X` gutter classes and
the `align-items-X` arrangement classes, we ensure the
positioning and spacing is the same in each module.

This allows to effectively remove the `form-inline-spaced`
class since this can easily be done with flex-form
and the mentioned classes in bootstrap v5.

Adjustments done:

* Use of flex-form to align and position form fields
* Add missing labels and unify arrangement
* Adjust size and position of form components
* Unify select field size
* Unify search field display
* Unify submit and reset button styles

Resolves: #94042
Resolves: #94026
Releases: master
Change-Id: Iae95a3e90179c6708ce6e1f0f272f267b6cae645
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/68926
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>