[FEATURE] Add filter option "Never hit" to Redirects module
The Redirects module is extended for a new filter option
"Never hit". When used, the list is filtered for redirects,
which were never hit before. This is especially handy
for editors managing large sites, while the "automatic
redirect generation on slug change" is enabled.

The redirect demand already features the property
"maxHits", which however was previously only used
for the CLI command and is now also used for the
new filter option.

Resolves: #94489
Releases: master
Change-Id: Ied510109de1c6bfb651ea0b8fdec0cfecbfab43a
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69823
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>