[FEATURE] Use LinkResult to generated Typolinks in Frontend
This change introduces a new immutable object "LinkResult" along
with an interface, containing the base result of a generated
link by TypoLink.

This object contains all information needed to put
together a <a> tag or return a pure URL.

For the time being this new class is used to build
links from the typolink link builders, and in addition
should be able to be returned fully by typolink
in the future.

In addition, this object helps to build links needed
for e.g. JSON responses to contain all information
of the link to be serialized.

Resolves: #94889
Releases: master
Change-Id: Ic9383a1a0f0ec93ba3aa352235003268e00a2f10
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/70590
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>