[BUGFIX] Fix notification for autogenerated slug redirects
Since #94114 EditDocumentController POST requests are
redirected to GET. (Inline) javascript that is added to
PageRenderer during processing of such POST request never
reaches to the browser and therefore caused notifications
to be omitted for automatically created redirects (#89115).

Use a update signal to delay notification delivery to the
GET request.

Also fix a misleading language label to clarifiy
that slugs of sub pages are being reverted, not the
slug that has been changed by the editor.

Releases: master
Resolves: #95123
Related: #94114
Related: #89115
Change-Id: Ic980ee8c18ec93655d38e3ab7626ac6b7a680ba0
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71410
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>