[BUGFIX] Catch UnknownLinkHandlerException in view helper
EXT:redirect is using the TargetPageIdViewHelper to
get the page ID from a "t3://page?uid=" link. If this link
is however somehow malformed, e.g. "t3://page=uid=",
this previously lead to an exception being thrown, making
the redirects module unusable for an editor. Since an
editor usually does not have access to pid=0, it needed
an administrator to make the module accessible again.

This is now fixed by catching this exception in the
view helper.

Resolves: #95488
Releases: master, 10.4
Change-Id: I27625f22a8b6f6a197402742c781d5fe8adae7e4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71500
Tested-by: core-ci <typo3@b13.com>
Tested-by: Sybille Peters <sypets@gmx.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Sybille Peters <sypets@gmx.de>
Reviewed-by: Benni Mack <benni@typo3.org>