[BUGFIX] Initialize language object in SlugServiceTest
Tests in SlugServiceTest from ext:redirects depend on
an initialized LanguageService, which has not been done
as there are no real frontend request executed, thus
the normal FE Bootstrap was not run. This leads to
undefined index errors because $GLOBALS['LANG'] is not
initalized.

The patch initalizes the LanguageService using the
Bootstrap::initalizeLanguageObject() in the setup().

Resolves: #95547
Releases: master
Change-Id: I53fb8877dd2692968c342e8df266f6268384ed1f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71517
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>