[BUGFIX] Ensure cyrilic work as regexp redirects capture group
Requests with special characters like for ex.
cyrilic characters in it would be rawurlencoded
by the used Uri() components. Webservers like
Apache normally decodeds them for the php process.

'RedirectService::matchRedirect()' has been
using rawurldecode() to decode the path to work
properly to resolve path with such characters.

RegExp redirects may have capture groups which
should be replaced in the target url for the
redirect. The corresponding replace method has
not been working with urldecoded variant and as
such the encoded value was used.

Furthermore if for ex. cyrilic characters where
in the path, and not only part of the capture
group the seconady preg_match in that replace
method did not matched the regexp which has
matched as redirect. Additional it could be
that cyrilic characters would be used for
the regexp in the query part, if respect
query parameters options is enabled for the
redirect.

This patch adds a rawurldecode() call to decode
the path and query, if needed,  to the method
`RedirectService::replaceRegExpCaptureGroup()`
and covers these cases with unit tests.

Releases: master, 10.4
Resolves: #94934
Change-Id: I464896d7efe2af20f5d2b725195b231f428760a3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71610
Tested-by: core-ci <typo3@b13.com>
Tested-by: Jochen <rothjochen@gmail.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Jochen <rothjochen@gmail.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>