[BUGFIX] Ensure regexp capture group works with relative target
RedirectService has been using LinkHandler to retrieve
the targeturl, thus supporting all TYPO3 has to offer,
like t3:// support, but also the legacy types for link
fields.

This had the effect, that for regexp redirects with
capture group replacement, this has not been executed,
this leading to incorrect redirects.

This patch assumes for undetermined urls for redirect
targets starting with an '/', that they are meant as
relative target redirects on the current request host.

This patch returns link details as type 'url' with the
relative path in the resolving method for redirects,
thus returning correct relative redirects but ensures
that regexp capture group replacements are done.

Redirects target without a leading slash could not
be handled in the same way, as they are assumed and
resolved as page links with the target as pageid.
These cannot be safely changed. Further tests are
added to cover these cases.

Relative redirects are executed relatively, without
adding a domain and scheme.

Resolves: #91796
Related: #89799
Releases: master, 10.4
Change-Id: I944ad4570b0cc8c3cdaf1a86c5383facb76c0860
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/71917
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>