[TASK] Remove unused language labels
In a previous patch, some language labels were unified in
EXT:redirects to that the same label was used for each
field in the list module as well as in the edit view.

Due to this some of the labels have become redundant and
are now removed.

Resolves: #96015
Related: #95832
Releases: main
Change-Id: I2fd9c650dcca0aea81999a3de2b427895984c70f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/72708
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>