[TASK] Use new backend templating in various controllers
This adapts a series of controllers and view
templates to use the ModuleTemplate as view.

The strategy is similar: $moduleTemplate is renamed
to $view, the former "body" view based on
BackendTemplateView is gone. Templates use the
backend Module.html as layout.

Some controllers are left out from this
patch since they need bigger refactoring.

Change-Id: Ie213840da46378c622fd35f797addc7b55f165fe
Resolves: #96745
Related: #96730
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/73180
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>