[BUGFIX] Ensure ext:redirects backend module pagination works
PostgresSQL is picky if aggregation functions like `order by`,
`group by` and similar are used, but missing the corresponding
columns in the "SELECT" column listing.

ext:redirects backend module builds query using the QueryBuilder
based on the demand dto object, which so far is a good practice.
To get the total count of records based on selection and properly
building the pagination it uses the same technique but executing
a simple count query, which leads to the aggregation error using
PostgresSQL dbms due the containing `order by` sortings.

This changes simply omits the orderBy part of the build query to
avoid this error and still counting on a correct full dataset.

Resolves: #98534
Releases: main, 11.5
Change-Id: I09e5f88d9d4f7c0aa578f8e1503cd1b17bc2c5dc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76052
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>