[FEATURE] Add TypoScript as Request attribute
The patch introduces the new request attribute
TypoScriptRequestAttribute as 'frontend.typoscript'.

It is set up in the Frontend middleware chain,
extensions and content objects can use this
attribute to retrieve TypoScript settings and
setup.

This obsoletes usages of TSFE->tmpl, the patch
switches usages to the request attribute.

Note we're also establishing a better wording:
The old TypoScript 'constants' are now called
'settings'. This matches much better, should avoid
confusion for newcomers and is in line with further
renaming of a vaguely anticipated new 'constants
editor' towards 'settings editing' or similar.
Further patches in this naming area will follow.

Change-Id: Ib6ffb91db9bf0976f39759b12983d78418d64efa
Resolves: #98914
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76239
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>