[FEATURE] Introduce new PSR-14 BeforeRedirectMatchDomainEvent
A new PSR-14 event `BeforeRedirectMatchDomainEvent` is introduced
to the RedirectService, allowing extension authors to implement a
custom redirect matching.

This can be used to implement custom logic to match the request
and determine a redirect. For example, dealing with `GET` request
params and changing the target based on the value of an parameter.

Therefore, special and rare use-cases can be implemented avoiding
complex all-scenario match logic in the core and keeping it to a
generic and overall cases in the core-implementation.

Resolves: #99803
Releases: main
Change-Id: I1819c3cd27b6a764000b61dc9cde580fa3b6d40d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77675
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>