[TASK] Streamline checkbox and radio rendering
We still carry around a lot of legacy checkbox
and radio styling, this patch reduces the
the complexity of existing overrides and drops
legacy classes like .checkbox and .radio and
their relatives.

We are adapting to the modern checkbox and
radio rendering at all places.

The .form-switch now drops all overrides and
takes full advantage of the bootstrap included
component.

The modifier .form-check-type-labeled-toggle
has been rebuilt to be accessible through
keyboard navigation and we are dropping
the custom green and red colorings for a
more unified primary color activation color.

The modifier .form-check-type-icon-toggle
has also been adapted, it allows providing
custom icons for the checked and unchecked states.
It is now also keyboard accessible and adapts
the primary activation color.

The modifier .form-check-type-toggle was adjusted
to be scoped. It is used in the list renderings and
the constant editor and has no label attached.

Resolves: #100537
Releases: main
Change-Id: I3dabe79a47ac232119f21f085e24347f28da4723
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78547
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>