[TASK] Use later DataHandler hook in `DataHandlerSlugUpdateHook`
Handling the pages slug changed event has been implemented with
a dedicated hook handling class, consuming two DataHandler hooks
to prepare and handle the redirect creation and slug update for
subpages. Auto redirect creation and subpage slug updating are
controlled by site setting configuration options.

The chosen second hook `processDatamap_postProcessFieldArray()`
is executed before the record is persisted into the database.

This change switches to `processDatamap_afterDatabaseOperations()`
as the second hook, so handling code can expect to have the
changes persisted to the database already. For example, this
would allow to use the slug generator in the `updateSlug()`
method in a dedicated change as possible solution for #96928.

Resolves: #101335
Related: #96928
Releases: main, 12.4
Change-Id: I5a83b6ecb9bebd7d1ebf1f3c704ca2ec2c10aa96
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80010
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>