[TASK] Use TCA and db defaults for auto-created redirect
This change gathers the default values for a redirect
from the definied TCA and uses the database schema as
fallback and use it for the auto-created redirects as
default before applying the SiteConfiguration setting,
manual defaults and dispatching the redirect event
`ModifyAutoCreateRedirectRecordBeforePersistingEvent`.

Reduces the need to implement a event listener when
the `sys_redirect` table is extended with not nullable
columns to ensure proper default value handling unknown
to the redirects slugservice.

Further superflous properties are removed in the
SlugService and remaining properties get native
types donated.

Resolves: #102938
Releases: main
Change-Id: I8cb406e3f571b9fdfe6f13bd356ea8d2e1668b2e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82656
Tested-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Guido Schmechel <guido.schmechel@brandung.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>