[TASK] Reduce `ext_tables.sql` for `ext:redirects`
It's a generic goal to provide database scheme defaults
for TCA defined tables with the on-going #101553 quest.

For `ext:redirects` the `target_statuscode` field has
not been removed due to the database default value.

Meanwhile it is ensured with #102938 that not only the
DataHandler but also the auto-create using QueryBuilder
respects TCA and database defaults - albeit overriden
with a manual default and based on the SiteConfiguration.

Therefore, this entry is now removed along with the todo
comment from the `ext_tables.sql`. Core uses proper API's
which ensures keeping the same default - whereas extension
not using proper api could be trapped with the removed
database default.

Note: Removing the `source_*` fields and `hitcount`
is not possible right now due to the simple fact that
`type=input` not having default schema creation yet.

Resolves: #102944
Related: #102938
Related: #101553
Releases: main
Change-Id: I841baf50cf6eb8ed170af9496e283641e4be917f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82661
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>