[FEATURE] Add TypoScript provider for sites and sets
TYPO3 sites are enhanced to be able to operate as TypoScript
templates. They act similar to sys_template records with
"clear" and "root" flags set. By design a site TypoScript provider
always defines a new scope (root-flag) and does not inherit from parent
sites (in the rootline). That means it behaves as if the clear-flag is
set in a sys_template record.
This behavior is not configurable by design, as TypoScript code sharing
is intended to be implemented via sharable sets (introduced in #103437).

TypoScript dependencies can be included via sets dependencies.
This mechanism supersedes the previous static_file_include's
or manual `@import` statements (they are still fine for local includes,
but should be avoided for cross-set/extensions dependencies),
as sets are automatically ordered and deduplicated.

Resolves: #103439
Releases: main
Change-Id: I971743fc551e51d945f45335dc6ad76404c6edba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83119
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>