[BUGFIX] Load full TypoScript setup in RedirectService
When a redirect is configured to use a `RecordLinkHandler`, the
redirect will fail with an exception, when the TypoScript setup
array is not available. This is the case, when at least one
regular page has been rendered.

With this change, it is ensured, that the full TypoScript
setup array is initialized in RedirectService, so later
creation of links in `DatabaseRecordLinkBuilder` will
have access to the full setup array via the
`frontend.typoscript` request attribute.

Resolves: #102892
Related: #105648
Releases: main, 13.4, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ide9dc87aa0a720bb059cce0fb6c0021e1fd68ba2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87871
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>