[BUGFIX] Make table sorting more usable
This patch improves the usability for sorting
columns in table listings. It provides clear
actionable buttons that state the action that is
done when pressing the button. That are placed
in a dropdown attached to the table header.

Actions that have previously cluttered the view
are now directly attached to the item in the dropdown
that also contains the sorting buttons. This helps
the user to identify better the effect of these
buttons, while also they contain readable labels for
all usergroups.

We now also allow to edit columns directly from
the list overview page.

Resolves: #105982
Related: #102542
Releases: main, 13.4
Change-Id: Ie3a528991ff21c96190306e28ea1f52cf9e3acc2
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87985
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>