[TASK] Add additional `EXT:redirects` functional tests
This change adds additional `RedirectService` functional
tests for `EXT:redirects` to cover some cases which has
been meanwhile fixed, but are still open in TYPO3 v12,
described in #99797.

Some test datasets are disabled in TYPO3 v12 and will be
enabled again with a following bugfix change for #99797.

One data set regarding cross site root redirect shows a
behaviour with a behaviour which needs to be decided if
that behaviour is correct. Added comments for that and
not in scope of this change to discuss or fix.

Resolves: #105987
Related: #99797
Releases: main, 13.4, 12.4
Change-Id: I503e0eeb95a7ba3633c542ab8c62d9141bf44989
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87878
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Torben Hansen <derhansen@gmail.com>
Reviewed-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <gh@faktor-e.de>
Reviewed-by: Torben Hansen <derhansen@gmail.com>
Tested-by: Garvin Hicking <gh@faktor-e.de>
Tested-by: Simon Schaufelberger <simonschaufi+typo3@gmail.com>