[TASK] Restrict Doctrine DBAL to minor version `4.2.*`
Doctrine DBAL 4 forced TYPO3 to extend a couple of classes
as replacement for the dropped event system to ensure the
same functionaly and required to copy some code which is
considered internal and partly made private by Doctrine.

That requires to adopt changes in copied code over time,
and due to the consideration as internal code this happens
also within minor version like upcoming `4.3.0` [1][2][3].

To mitigate this for the TYPO3 user base the dependency is
restricted to minor version ranges to coordinate doctrine
updates with eventually required code adoptions and mitigate
breaking composer instances while keeping bugfix updates.

The TYPO3 Core Team strifes to align with the Doctrine Team
to find patterns that allow for more seamless integration in
frameworks like TYPO3 and reduce the burdon of copied code.

Used command(s):

BIN_COMPOSER="$( which composer )" \
&& ${BIN_COMPOSER} require \
  -d typo3/sysext/core --no-update --no-install \
  'doctrine/dbal':'4.2.*' \
&& ${BIN_COMPOSER} require \
  -d typo3/sysext/install --no-update --no-install \
  'doctrine/dbal':'4.2.*' \
&& ${BIN_COMPOSER} require \
  -d typo3/sysext/redirects --no-update --no-install \
  'doctrine/dbal':'4.2.*' \
&& ${BIN_COMPOSER} require -w \
  'doctrine/dbal':'4.2.*'

[1] https://github.com/doctrine/dbal/pull/6932
[2] https://github.com/doctrine/dbal/pull/6933
[3] https://github.com/doctrine/dbal/pull/6944

Resolves: #106803
Releases: main, 13.4
Change-Id: I14a3b6f65561495e0406b3ac361633e1b770dbbc
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/89648
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Garvin Hicking <garvin@hick.ing>